+++ tests/test_dcos_e2e/backends/docker/test_docker.py
@@ -8,12 +8,13 @@ sibling modules.
 import subprocess
 import uuid
 from pathlib import Path
-from typing import Dict
+from typing import Iterator
 
 # See https://github.com/PyCQA/pylint/issues/1536 for details on why the errors
 # are disabled.
 import docker
 import pytest
+from docker.models.networks import Network
 from docker.types import Mount
 from py.path import local  # pylint: disable=no-name-in-module, import-error
 from requests_mock import Mocker, NoMockAddress
@@ -23,7 +24,7 @@ from dcos_e2e.backends import Docker
 from dcos_e2e.cluster import Cluster
 from dcos_e2e.docker_storage_drivers import DockerStorageDriver
 from dcos_e2e.docker_versions import DockerVersion
-from dcos_e2e.node import Node
+from dcos_e2e.node import Node, Transport
 
 
 @retry(
@@ -37,6 +38,23 @@ def _wait_for_docker(node: Node) -> None:
     node.
     """
     node.run(args=['docker', 'info'])
+
+
+def _get_container_from_node(node: Node) -> docker.models.containers.Container:
+    """
+    Return the container which represents the given ``node``.
+    """
+    client = docker.from_env(version='auto')
+    containers = client.containers.list()
+    matching_containers = []
+    for container in containers:
+        networks = container.attrs['NetworkSettings']['Networks']
+        for net in networks:
+            if networks[net]['IPAddress'] == str(node.public_ip_address):
+                matching_containers.append(container)
+
+    assert len(matching_containers) == 1
+    return matching_containers[0]
 
 
 class TestDockerBackend:
@@ -309,19 +327,6 @@ class TestLabels:
     Tests for setting labels on Docker containers.
     """
 
-    def _get_labels(self, node: Node) -> Dict[str, str]:
-        """
-        Return the labels on the container which maps to ``node``.
-        """
-        client = docker.from_env(version='auto')
-        containers = client.containers.list()
-        [container] = [
-            container for container in containers
-            if container.attrs['NetworkSettings']['IPAddress'] ==
-            str(node.public_ip_address)
-        ]
-        return dict(container.labels)
-
     def test_custom(self) -> None:
         """
         It is possible to set node Docker container labels.
@@ -351,61 +356,144 @@ class TestLabels:
 
         with Cluster(cluster_backend=cluster_backend) as cluster:
             for node in cluster.masters:
-                node_labels = self._get_labels(node=node)
+                node_labels = dict(_get_container_from_node(node=node).labels)
                 assert node_labels[cluster_key] == cluster_value
                 assert node_labels[master_key] == master_value
                 assert agent_key not in node_labels
                 assert public_agent_key not in node_labels
 
             for node in cluster.agents:
-                node_labels = self._get_labels(node=node)
+                node_labels = dict(_get_container_from_node(node=node).labels)
                 assert node_labels[cluster_key] == cluster_value
                 assert node_labels[agent_key] == agent_value
                 assert master_key not in node_labels
                 assert public_agent_key not in node_labels
 
             for node in cluster.public_agents:
-                node_labels = self._get_labels(node=node)
+                node_labels = dict(_get_container_from_node(node=node).labels)
                 assert node_labels[cluster_key] == cluster_value
                 assert node_labels[public_agent_key] == public_agent_value
                 assert master_key not in node_labels
                 assert agent_key not in node_labels
 
 
-class TestEtcHosts:
+class TestNetworks:
     """
-    Test the creation of ``/etc/hosts``.
+    Tests for Docker container networks.
+
+    On macOS, by default, it is not possible to SSH to containers without
+    forwarded ports.
+
+    We therefore recommend that people use ``dcos-docker setup-mac-network``.
+    This makes it possible to SSH to containers in the default network range.
+
+    However, in these tests, we use custom networks.
+    Using the VPN created by ``dcos-docker setup-mac-network`` it is not
+    possible to SSH to containers on custom networks.
+    See https://github.com/wojas/docker-mac-network#openvpn for details.
+
+    Therefore, we use the ``Transport.DOCKER_EXEC`` transport to communicate
+    with nodes.
     """
 
-    def test_registry_hosts(self) -> None:
+    @pytest.fixture()
+    def docker_network(self) -> Iterator[Network]:
         """
-        One master has 127.0.0.1 in its ``/etc/hosts`` file mapping to
-        ``registry.local``. All other nodes have that master's IP address
-        mapping to ``registry.local``.
+        Return a Docker network.
         """
-        with Cluster(cluster_backend=Docker()) as cluster:
-            nodes = {*cluster.masters, *cluster.agents, *cluster.public_agents}
+        client = docker.from_env(version='auto')
+        ipam_pool = docker.types.IPAMPool(
+            subnet='172.28.0.0/16',
+            iprange='172.28.0.0/24',
+            gateway='172.28.0.254',
+        )
+        network = client.networks.create(
+            # The container name prefix "dcos-e2e-" matches the prefix used in
+            # the "clean" Makefile target.
+            name='dcos-e2e-network-{random}'.format(random=uuid.uuid4()),
+            driver='bridge',
+            ipam=docker.types.IPAMConfig(pool_configs=[ipam_pool]),
+            attachable=False,
+        )
+        try:
+            yield network
+        finally:
+            network.remove()
 
-            registries = {}
+    def test_custom_docker_network(
+        self,
+        docker_network: Network,
+    ) -> None:
+        """
+        When a network is specified on the Docker backend, each container is
+        connected to the default bridge network ``docker0`` and in addition it
+        also connected to the custom network.
 
-            for node in nodes:
-                etc_hosts = node.run(args=['cat', '/etc/hosts'])
-                etc_hosts_contents = etc_hosts.stdout.strip().decode()
-                for line in etc_hosts_contents.split('\n'):
-                    split_line = line.split()
-                    address = split_line[0]
-                    hosts = split_line[1:]
-                    if hosts == ['registry.local']:
-                        registries[node] = address
+        The ``Node``'s IP addresses correspond to the custom network.
+        """
+        with Cluster(
+            cluster_backend=Docker(
+                network=docker_network,
+                transport=Transport.DOCKER_EXEC,
+            ),
+            agents=0,
+            public_agents=0,
+        ) as cluster:
+            (master, ) = cluster.masters
+            container = _get_container_from_node(master)
+            networks = container.attrs['NetworkSettings']['Networks']
+            assert networks.keys() == set(['bridge', docker_network.name])
+            custom_network_ip = networks[docker_network.name]['IPAddress']
+            assert custom_network_ip == str(master.public_ip_address)
+            assert custom_network_ip == str(master.private_ip_address)
 
-            [registry_node] = [
-                node for node, registry_address in registries.items()
-                if registry_address == '127.0.0.1'
-            ]
+    def test_docker_exec_transport(
+        self,
+        docker_network: Network,
+        tmpdir: local,
+    ) -> None:
+        """
+        ``Node`` operations with the Docker exec transport work even if the
+        node is on a custom network.
+        """
+        with Cluster(
+            cluster_backend=Docker(
+                network=docker_network,
+                transport=Transport.DOCKER_EXEC,
+            ),
+            agents=0,
+            public_agents=0,
+        ) as cluster:
+            (master, ) = cluster.masters
+            content = str(uuid.uuid4())
+            local_file = tmpdir.join('example_file.txt')
+            local_file.write(content)
+            random = uuid.uuid4().hex
+            master_destination_dir = '/etc/{random}'.format(random=random)
+            master_destination_path = Path(master_destination_dir) / 'file.txt'
+            master.send_file(
+                local_path=Path(str(local_file)),
+                remote_path=master_destination_path,
+                transport=Transport.DOCKER_EXEC,
+            )
+            args = ['cat', str(master_destination_path)]
+            result = master.run(args=args, transport=Transport.DOCKER_EXEC)
+            assert result.stdout.decode() == content
 
-            assert registry_node in cluster.masters
-
-            assert set(registries.values()) == {
-                '127.0.0.1',
-                str(registry_node.private_ip_address),
-            }
+    def test_default(self) -> None:
+        """
+        By default, the only network a container is in is the Docker default
+        bridge network.
+        """
+        with Cluster(
+            cluster_backend=Docker(),
+            agents=0,
+            public_agents=0,
+        ) as cluster:
+            (master, ) = cluster.masters
+            container = _get_container_from_node(master)
+            networks = container.attrs['NetworkSettings']['Networks']
+            assert networks.keys() == set(['bridge'])
+            bridge_ip_address = networks['bridge']['IPAddress']
+            assert bridge_ip_address == str(master.public_ip_address)
+            assert bridge_ip_address == str(master.private_ip_address)